# backend/swagger.yaml
openapi: 3.0.3
info:
  title: ERP System Backend API
  description: |
    Complete API documentation for the ERP System Backend microservices.
    
    ## Services
    - **Auth Service**: Authentication and user management
    - **Organization Service**: Organizations, companies, and locations
    - **Inventory Service**: Products and stock management
    - **License Service**: Application licenses and assignments
    - **Subscription Service**: Subscription plans and billing
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    
  version: 1.0.0
  contact:
    name: ERP System Team
    email: support@erpsystem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Auth Service (Development)
  - url: http://localhost:8082/api/v1
    description: Organization Service (Development)
  - url: http://localhost:8083/api/v1
    description: Inventory Service (Development)
  - url: http://localhost:8084/api/v1
    description: License Service (Development)
  - url: http://localhost:8085/api/v1
    description: Subscription Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          type: object
...
    BaseModel:
      type: object
      properties:
        id:
          type: string
          format: objectid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: objectid
        updated_by:
          type: string
          format: objectid
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        deleted_by:
          type: string
          format: objectid
          nullable: true
...
    SubscriptionPlan:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            name:
              type: string
            code:
              type: string
            description:
              type: string
            tier:
              type: string
              enum: [free, starter, professional, enterprise]
            price_monthly:
              type: number
              format: float
            price_yearly:
              type: number
              format: float
            features:
              type: object
            is_active:
              type: boolean
            is_public:
              type: boolean
    Subscription:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            organization_id:
              type: string
              format: objectid
            plan_id:
              type: string
              format: objectid
            status:
              type: string
              enum: [trial, active, past_due, cancelled, expired]
            billing_cycle:
              type: string
              enum: [monthly, yearly]
            current_period_start:
              type: string
              format: date-time
            current_period_end:
              type: string
              format: date-time
            cancel_at_period_end:
              type: boolean
paths:
...
  /plans:
    get:
      tags:
        - Plans
      summary: List subscription plans
      description: Get list of subscription plans (public endpoint)
      parameters:
        - name: tier
          in: query
          schema:
            type: string
            enum: [free, starter, professional, enterprise]
        - name: public
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Plans retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    
    post:
      tags:
        - Plans
      summary: Create plan
      description: Create a new subscription plan (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - code
                - tier
              properties:
                name:
                  type: string
                code:
                  type: string
                description:
                  type: string
                tier:
                  type: string
                  enum: [free, starter, professional, enterprise]
                price_monthly:
                  type: number
                price_yearly:
                  type: number
                is_public:
                  type: boolean
      responses:
        '201':
          description: Plan created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionPlan'
  /plans/{id}:
    get:
      tags:
        - Plans
      summary: Get plan
      description: Retrieve subscription plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plan retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionPlan'
  /organizations/{org_id}/subscription:
    get:
      tags:
        - Subscriptions
      summary: Get subscription
      description: Get organization's current subscription
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subscription retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Subscription'
    
    post:
      tags:
        - Subscriptions
      summary: Subscribe
      description: Create a subscription for an organization
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - plan_id
                - billing_cycle
              properties:
                plan_id:
                  type: string
                billing_cycle:
                  type: string
                  enum: [monthly, yearly]
                start_trial:
                  type: boolean
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Subscription'
  /organizations/{org_id}/subscription/cancel:
    post:
      tags:
        - Subscriptions
      summary: Cancel subscription
      description: Cancel organization's subscription
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                immediate:
                  type: boolean
                  description: Cancel immediately or at period end
                reason:
                  type: string
      responses:
        '200':
          description: Subscription cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'